personal-chef/requirements.txt
annotated-types==0.7.0
anyio==4.8.0
art==6.4
certifi==2025.1.31
charset-normalizer==3.4.1
colorama==0.4.6
distro==1.9.0
fire==0.7.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
jiter==0.8.2
markdown-it-py==3.0.0
mdurl==0.1.2
numpy==2.0.2
openai==1.65.4
pillow==11.1.0
pydantic==2.10.6
pydantic_core==2.27.2
pyfiglet==1.0.2
Pygments==2.19.1
python-dotenv==1.0.1
requests==2.32.3
rich==13.9.4
sniffio==1.3.1
termcolor==2.5.0
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.3.0
